@@ -3,47 +3,45 @@
   "version": "0.1.0",
   "private": true,
   "engines": {
-    "node": "10.19.0"
+    "node": "18.4.0"
   },
   "scripts": {
     "serve": "vue-cli-service serve --open",
     "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint",
-    "postinstall": "yarn build"
+    "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "axios": "^0.18.0",
-    "vue": "^2.5.13",
-    "vue-router": "^3.0.1",
-    "vuex": "^3.0.1"
+    "axios": "^0.27.2",
+    "bootstrap-vue-3": "^0.1.13",
+    "prettier": "^2.7.1",
+    "vue": "^3.2.37",
+    "vue-router": "^4.0.16",
+    "vuex": "^4.0.2",
+    "vuex-persistedstate": "^4.1.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^3.0.0-beta.6",
-    "@vue/cli-plugin-eslint": "^3.0.0-beta.6",
-    "@vue/cli-service": "^3.0.0-beta.6",
-    "@vue/eslint-config-standard": "^3.0.0-beta.6",
-    "bootstrap": "^4.3.1",
-    "bootstrap-vue": "2.16.0",
-    "highcharts": "8.2.0",
-    "highcharts-vue": "1.3.5",
-    "lint-staged": "^6.0.0",
-    "moment": "2.27.0",
+    "@vue/cli": "^5.0.6",
+    "@vue/cli-plugin-babel": "~5.0.6",
+    "@vue/cli-plugin-eslint": "~5.0.6",
+    "@vue/cli-plugin-router": "~5.0.6",
+    "@vue/cli-plugin-vuex": "~5.0.6",
+    "@vue/cli-service": "^5.0.6",
+    "@vue/compiler-sfc": "^3.2.37",
+    "bootstrap": "^5.1.3",
+    "eslint": "^7.32.0",
+    "eslint-plugin-vue": "^7.0.0-beta.4",
+    "highcharts": "10.1.0",
+    "highcharts-vue": "1.4.0",
+    "lint-staged": "^13.0.3",
+    "moment": "2.29.3",
     "moment-timezone": "0.5.34",
-    "node-sass": "^4.7.2",
-    "sass-loader": "^6.0.6",
-    "vue-template-compiler": "^2.5.13",
-    "vuex-persistedstate": "3.1.0"
+    "node-sass": "^7.0.1",
+    "sass-loader": "^13.0.2",
+    "vue-template-compiler": "^2.7.0"
   },
   "babel": {
     "presets": [
       "@vue/app"
-    ]
-  },
-  "eslintConfig": {
-    "root": true,
-    "extends": [
-      "plugin:vue/essential",
-      "@vue/standard"
     ]
   },
   "postcss": {
@@ -56,22 +54,9 @@
     "last 2 versions",
     "not ie <= 8"
   ],
-  "gitHooks": {
-    "pre-commit": "lint-staged"
-  },
   "prettier": {
     "singleQuote": true,
     "trailingComma": "all",
     "semi": false
-  },
-  "lint-staged": {
-    "*.js": [
-      "vue-cli-service lint",
-      "git add"
-    ],
-    "*.vue": [
-      "vue-cli-service lint",
-      "git add"
-    ]
   }
 }
