@@ -1,10 +1,11 @@
 import Vue from 'vue'
+import { createApp } from "vue";
 import App from './App.vue'
-import router from '@/router'
-import store from '@/store'
+import router from './router'
+import store from './store'
 
 // bootstrap mobile first
-import BootstrapVue from 'bootstrap-vue'
+import BootstrapVue from 'bootstrap-vue-3'
 import 'bootstrap/dist/css/bootstrap.css'
 import 'bootstrap-vue/dist/bootstrap-vue.css'
 
@@ -15,16 +16,9 @@ import stockInit from 'highcharts/modules/stock'
 
 stockInit(Highcharts)
 
-Vue.use(HighchartsVue)
-
-Vue.use(BootstrapVue)
-
-Vue.config.devtools = true
-
-Vue.config.productionTip = false
-
-new Vue({
-  router,
-  store,
-  render: h => h(App)
-}).$mount('#app')
+const app = createApp(App);
+app.use(store)
+app.use(router)
+app.use(HighchartsVue)
+app.use(BootstrapVue)
+app.mount("#app");
